Vis statisk tekst om andre forelder dersom man har samme forelder som en annen på eøs
