+++ src/frontend/components/SøknadsSteg/EøsSteg/Barn/EøsForBarn.tsx
@@ -4,14 +4,18 @@ import { useIntl } from 'react-intl';
 
 import { ESvar } from '@navikt/familie-form-elements';
 
+import { useApp } from '../../../../context/AppContext';
+import { barnDataKeySpørsmål, IBarnMedISøknad } from '../../../../typer/barn';
 import { BarnetsId } from '../../../../typer/common';
 import { barnetsNavnValue, skalSkjuleAndreForelderFelt } from '../../../../utils/barn';
 import { Arbeidsperiode } from '../../../Felleskomponenter/Arbeidsperiode/Arbeidsperiode';
+import KomponentGruppe from '../../../Felleskomponenter/KomponentGruppe/KomponentGruppe';
 import { Pensjonsperiode } from '../../../Felleskomponenter/Pensjonsmodal/Pensjonsperiode';
 import SkjemaFieldset from '../../../Felleskomponenter/SkjemaFieldset';
 import SpråkTekst from '../../../Felleskomponenter/SpråkTekst/SpråkTekst';
 import Steg from '../../../Felleskomponenter/Steg/Steg';
 import { Utbetalingsperiode } from '../../../Felleskomponenter/UtbetalingerModal/Utbetalingsperiode';
+import EøsAndreForelderOppsummering from '../../Oppsummering/OppsummeringSteg/Eøs/EøsAndreForelderOppsummering';
 import { useEøsForBarn } from './useEøsForBarn';
 
 const EøsForBarn: React.FC<{ barnetsId: BarnetsId }> = ({ barnetsId }) => {
@@ -29,6 +33,16 @@ const EøsForBarn: React.FC<{ barnetsId: BarnetsId }> = ({ barnetsId }) => {
         fjernArbeidsperiode,
     } = useEøsForBarn(barnetsId);
     const intl = useIntl();
+    const { søknad } = useApp();
+
+    const andreBarnSomErFyltUt = søknad.barnInkludertISøknaden.filter(
+        barnISøknad =>
+            barnISøknad.barnErFyltUt && barnISøknad.id !== barn.id && !!barnISøknad.andreForelder
+    );
+
+    const barnMedSammeForelder: IBarnMedISøknad | undefined = andreBarnSomErFyltUt.find(
+        annetBarn => annetBarn.id === barn[barnDataKeySpørsmål.sammeForelderSomAnnetBarnMedId].svar
+    );
 
     return (
         <Steg
@@ -47,45 +61,60 @@ const EøsForBarn: React.FC<{ barnetsId: BarnetsId }> = ({ barnetsId }) => {
         >
             {!skalSkjuleAndreForelderFelt(barn) && (
                 <SkjemaFieldset tittelId={'ombarnet.andre-forelder'}>
-                    <Arbeidsperiode
-                        skjema={skjema}
-                        leggTilArbeidsperiode={leggTilArbeidsperiode}
-                        fjernArbeidsperiode={fjernArbeidsperiode}
-                        arbeiderEllerArbeidetFelt={skjema.felter.andreForelderArbeidNorge}
-                        registrerteArbeidsperioder={skjema.felter.andreForelderArbeidsperioderNorge}
-                        andreForelderData={{
-                            erDød: barn.andreForelderErDød.svar === ESvar.JA,
-                            barn: barn,
-                        }}
-                    />
-                    <Pensjonsperiode
-                        skjema={skjema}
-                        mottarEllerMottattPensjonFelt={skjema.felter.andreForelderPensjonNorge}
-                        leggTilPensjonsperiode={leggTilPensjonsperiode}
-                        fjernPensjonsperiode={fjernPensjonsperiode}
-                        andreForelderData={{
-                            erDød: barn.andreForelderErDød.svar === ESvar.JA,
-                            barn: barn,
-                        }}
-                        registrertePensjonsperioder={
-                            skjema.felter.andreForelderPensjonsperioderNorge
-                        }
-                    />
-                    <Utbetalingsperiode
-                        skjema={skjema}
-                        mottarEllerMottattUtbetalingFelt={
-                            skjema.felter.andreForelderAndreUtbetalinger
-                        }
-                        leggTilUtbetalingsperiode={leggTilAndreUtbetalingsperiode}
-                        fjernUtbetalingsperiode={fjernAndreUtbetalingsperiode}
-                        andreForelderData={{
-                            erDød: barn.andreForelderErDød.svar === ESvar.JA,
-                            barn: barn,
-                        }}
-                        registrerteUtbetalingsperioder={
-                            skjema.felter.andreForelderAndreUtbetalingsperioder
-                        }
-                    />
+                    {!barnMedSammeForelder ? (
+                        <KomponentGruppe>
+                            <Arbeidsperiode
+                                skjema={skjema}
+                                leggTilArbeidsperiode={leggTilArbeidsperiode}
+                                fjernArbeidsperiode={fjernArbeidsperiode}
+                                arbeiderEllerArbeidetFelt={skjema.felter.andreForelderArbeidNorge}
+                                registrerteArbeidsperioder={
+                                    skjema.felter.andreForelderArbeidsperioderNorge
+                                }
+                                andreForelderData={{
+                                    erDød: barn.andreForelderErDød.svar === ESvar.JA,
+                                    barn: barn,
+                                }}
+                            />
+                            <Pensjonsperiode
+                                skjema={skjema}
+                                mottarEllerMottattPensjonFelt={
+                                    skjema.felter.andreForelderPensjonNorge
+                                }
+                                leggTilPensjonsperiode={leggTilPensjonsperiode}
+                                fjernPensjonsperiode={fjernPensjonsperiode}
+                                andreForelderData={{
+                                    erDød: barn.andreForelderErDød.svar === ESvar.JA,
+                                    barn: barn,
+                                }}
+                                registrertePensjonsperioder={
+                                    skjema.felter.andreForelderPensjonsperioderNorge
+                                }
+                            />
+                            <Utbetalingsperiode
+                                skjema={skjema}
+                                mottarEllerMottattUtbetalingFelt={
+                                    skjema.felter.andreForelderAndreUtbetalinger
+                                }
+                                leggTilUtbetalingsperiode={leggTilAndreUtbetalingsperiode}
+                                fjernUtbetalingsperiode={fjernAndreUtbetalingsperiode}
+                                andreForelderData={{
+                                    erDød: barn.andreForelderErDød.svar === ESvar.JA,
+                                    barn: barn,
+                                }}
+                                registrerteUtbetalingsperioder={
+                                    skjema.felter.andreForelderAndreUtbetalingsperioder
+                                }
+                            />
+                        </KomponentGruppe>
+                    ) : (
+                        barnMedSammeForelder?.andreForelder && (
+                            <EøsAndreForelderOppsummering
+                                barn={barn}
+                                andreForelder={barnMedSammeForelder.andreForelder}
+                            />
+                        )
+                    )}
                 </SkjemaFieldset>
             )}
         </Steg>
