+++ src/frontend/components/SøknadsSteg/Oppsummering/Oppsummering.tsx
@@ -14,8 +14,8 @@ import { RouteEnum } from '../../../typer/routes';
 import SpråkTekst from '../../Felleskomponenter/SpråkTekst/SpråkTekst';
 import Steg from '../../Felleskomponenter/Steg/Steg';
 import DinLivssituasjonOppsummering from './OppsummeringSteg/DinLivssituasjonOppsummering';
-import EøsBarnOppsummering from './OppsummeringSteg/EøsBarnOppsummering';
-import EøsSøkerOppsummering from './OppsummeringSteg/EøsSøkerOppsummering';
+import EøsBarnOppsummering from './OppsummeringSteg/Eøs/EøsBarnOppsummering';
+import EøsSøkerOppsummering from './OppsummeringSteg/Eøs/EøsSøkerOppsummering';
 import OmBarnaOppsummering from './OppsummeringSteg/OmBarnaOppsummering';
 import OmBarnetOppsummering from './OppsummeringSteg/OmBarnet/OmBarnetOppsummering';
 import OmDegOppsummering from './OppsummeringSteg/OmDegOppsummering';
